+++ src/browser.ts
@@ -1,5 +1,11 @@
-import { rant as run, createRant, enUS } from "./rant.ts";
+import {
+  compile,
+  createRant,
+  enUS,
+  explain,
+  rant as run,
+} from "./index.ts";
 
-const api = Object.assign(run, { createRant, enUS, rant: run });
+const api = Object.assign(run, { compile, createRant, explain, enUS, rant: run });
 
 export default api;